Refactor claim repository to execute postProcess on listAll and move code that doesn't belong to this module to either util or claim review repository. Refactor personality repository too.
