+++ server/api/repository/claimReview.ts
@@ -1,5 +1,6 @@
 import { ILogger } from "../../lib/loggerInterface";
 
+const util = require("../../lib/util");
 const ClaimReview = require("../model/claimReviewModel");
 const Claim = require("../model/claimModel");
 const Source = require("../model/sourceModel");
@@ -56,6 +57,67 @@ export default class ClaimReviewRepository {
         return ClaimReview.findById(claimReviewId)
             .populate("claims", "_id title")
             .populate("sources", "_id link classification");
+    }
+
+    getReviewsByClaimId(claimId) {
+        return ClaimReview.aggregate([
+            { $match: { claim: claimId } },
+            {
+                $group: {
+                    _id: "$sentence_hash",
+                    topClassification: {
+                        $accumulator: {
+                            init: function() {
+                                return {};
+                            },
+                            accumulate: function(state, classification) {
+                                if (!state[classification]) {
+                                    state[classification] = 1;
+                                } else {
+                                    state[classification]++;
+                                }
+
+                                return state;
+                            },
+                            accumulateArgs: ["$classification"],
+                            merge: function(state1, state2) {
+                                return { ...state1, ...state2 };
+                            },
+                            finalize: function(state) {
+                                // Find the classification with bigger count
+                                const topClassification = Object.keys(
+                                    state
+                                ).reduce((acc, classification) => {
+                                    if (!state[acc]) {
+                                        return classification;
+                                    } else {
+                                        return state[acc] >=
+                                            state[classification]
+                                            ? acc
+                                            : classification;
+                                    }
+                                }, "");
+                                // TODO: what can we do about ties?
+                                return {
+                                    classification: topClassification,
+                                    count: state[topClassification]
+                                };
+                            },
+                            lang: "js"
+                        }
+                    }
+                }
+            }
+        ]).option({ serializeFunctions: true });
+    }
+
+    async getReviewStatsByClaimId(claimId) {
+        const reviews = await ClaimReview.aggregate([
+            { $match: { claim: claimId } },
+            { $group: { _id: "$classification", count: { $sum: 1 } } },
+            { $sort: { count: -1 } }
+        ]);
+        return util.formatStats(reviews);
     }
 
     async update(claimReviewId, claimReviewBody) {
