+++ server/lib/util.ts
@@ -32,7 +32,30 @@ function formatStats(reviews, slice = false) {
     });
     return { total, reviews: slice ? result.slice(0, 3) : result };
 }
+
+/**
+ * https://medium.com/javascript-in-plain-english/javascript-merge-duplicate-objects-in-array-of-objects-9a76c3a1c35c
+ * @param array
+ * @param property
+ */
+function mergeObjectsInUnique<T>(array: T[], property: any): T[] {
+    const newArray = new Map();
+
+    array.forEach((item: T) => {
+        const propertyValue = item[property];
+        newArray.has(propertyValue)
+            ? newArray.set(propertyValue, {
+                  ...item,
+                  ...newArray.get(propertyValue)
+              })
+            : newArray.set(propertyValue, item);
+    });
+
+    return Array.from(newArray.values());
+}
+
 module.exports = {
     router: createRouter,
-    formatStats
+    formatStats,
+    mergeObjectsInUnique
 };
